+++ digiplan/map/models.py
@@ -3,7 +3,7 @@
 
 import pandas as pd
 from django.contrib.gis.db import models
-from django.db.models import Sum
+from django.db.models import Count, Sum
 from django.utils.translation import gettext_lazy as _
 
 from .managers import LabelMVTManager, RegionMVTManager, StaticMVTManager
@@ -708,3 +708,150 @@ class PotentialAreaWindSTP2018EG(StaticRegionModel):  # noqa: D101
 class PotentialAreaWindSTP2024VR(StaticRegionModel):  # noqa: D101
     data_file = "potentialarea_wind_stp_2024_vr"
     layer = "potentialarea_wind_stp_2024_vr"
+
+
+class PVgroundAreas(StaticRegionModel):
+    """Model holding PV on ground (dataset by RPG with areas)."""
+
+    name = models.CharField(max_length=255, null=True)
+    plan_type = models.CharField(max_length=255, null=True)
+    plan_status = models.CharField(max_length=255, null=True)
+    status = models.CharField(max_length=10, null=True)
+    capacity_net = models.FloatField(null=True)
+    year = models.BigIntegerField(null=True)
+    construction_start_date = models.CharField(max_length=10, null=True)
+    construction_end_date = models.CharField(max_length=10, null=True)
+
+    mun_id = models.ForeignKey(Municipality, on_delete=models.DO_NOTHING, null=True)
+
+    mapping = {
+        "geom": "MULTIPOLYGON",
+        "name": "name",
+        "plan_type": "plan_type",
+        "plan_status": "plan_status",
+        "status": "status",
+        "capacity_net": "capacity_net",
+        "year": "year",
+        "construction_start_date": "construction_start_date",
+        "construction_end_date": "construction_end_date",
+        "mun_id": {"id": "municipality_id"},
+    }
+
+    class Meta:  # noqa: D106
+        abstract = True
+
+
+class PVgroundAreasApproved(PVgroundAreas):
+    """Model holding PV on ground (dataset by RPG with areas): Approved units."""
+
+    data_file = "rpg_ols_pv_ground_approved"
+    layer = "rpg_ols_pv_ground_approved"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Freiflächen-PV (genehmigt)")
+        verbose_name_plural = _("Freiflächen-PV (genehmigt)")
+
+
+class PVgroundAreasOperating(PVgroundAreas):
+    """Model holding PV on ground (dataset by RPG with areas): Operating units."""
+
+    data_file = "rpg_ols_pv_ground_operating"
+    layer = "rpg_ols_pv_ground_operating"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Freiflächen-PV (in Betrieb)")
+        verbose_name_plural = _("Freiflächen-PV (in Betrieb)")
+
+
+class PVgroundAreasPlanned(PVgroundAreas):
+    """Model holding PV on ground (dataset by RPG with areas): Planned units."""
+
+    data_file = "rpg_ols_pv_ground_planned"
+    layer = "rpg_ols_pv_ground_planned"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Freiflächen-PV (geplant)")
+        verbose_name_plural = _("Freiflächen-PV (geplant)")
+
+
+class WindTurbine2(models.Model):
+    """Model holding wind turbines."""
+
+    name = models.CharField(max_length=255, null=True)
+    operator = models.CharField(max_length=255, null=True)
+    city = models.CharField(max_length=50, null=True)
+    zip_code = models.CharField(max_length=50, null=True)
+    commissioning_date = models.CharField(max_length=50, null=True)
+    capacity_net = models.FloatField(null=True)
+    hub_height = models.FloatField(null=True)
+    rotor_diameter = models.FloatField(null=True)
+    status = models.CharField(max_length=50, null=True)
+    geom = models.PointField(srid=4326)
+
+    mun_id = models.ForeignKey(Municipality, on_delete=models.DO_NOTHING, null=True)
+
+    objects = models.Manager()
+    vector_tiles = StaticMVTManager(columns=[])
+
+    mapping = {
+        "geom": "POINT",
+        "name": "name",
+        "operator": "operator",
+        "city": "city",
+        "zip_code": "zip_code",
+        "commissioning_date": "commissioning_date",
+        "capacity_net": "capacity_net",
+        "hub_height": "hub_height",
+        "rotor_diameter": "rotor_diameter",
+        "mun_id": {"id": "municipality_id"},
+    }
+
+    class Meta:  # noqa: D106
+        abstract = True
+
+    @classmethod
+    def quantity_per_municipality(cls) -> pd.DataFrame:
+        """
+        Calculate number of wind turbines per municipality.
+
+        Returns
+        -------
+        dpd.DataFrame
+            wind turbines per municipality
+        """
+        queryset = cls.objects.values("mun_id").annotate(units=Count("name")).values("mun_id", "units")
+        wind_turbines = pd.DataFrame.from_records(queryset).set_index("mun_id")
+        return wind_turbines["units"].reindex(Municipality.objects.all().values_list("id", flat=True), fill_value=0)
+
+
+class WindTurbine2Approved(WindTurbine2):
+    """Model holding PV on ground (dataset by RPG with areas): Approved units."""
+
+    data_file = "rpg_ols_wind_approved"
+    layer = "rpg_ols_wind_approved"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Windenergieanlage (genehmigt)")
+        verbose_name_plural = _("Windenergieanlagen (genehmigt)")
+
+
+class WindTurbine2Operating(WindTurbine2):
+    """Model holding PV on ground (dataset by RPG with areas): Operating units."""
+
+    data_file = "rpg_ols_wind_operating"
+    layer = "rpg_ols_wind_operating"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Windenergieanlage (in Betrieb)")
+        verbose_name_plural = _("Windenergieanlagen (in Betrieb)")
+
+
+class WindTurbine2Planned(WindTurbine2):
+    """Model holding PV on ground (dataset by RPG with areas): Planned units."""
+
+    data_file = "rpg_ols_wind_planned"
+    layer = "rpg_ols_wind_planned"
+
+    class Meta:  # noqa: D106
+        verbose_name = _("Windenergieanlage (geplant)")
+        verbose_name_plural = _("Windenergieanlagen (geplant)")